requirements.txt
# LangGraph / LangChain core
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0

# Telegram bot
python-telegram-bot>=21.0

# Gmail / Google APIs
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.150.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.8.0
aiohttp>=3.10.0

# Research workflow
duckduckgo-search>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
ddgs>=1.0.0
